residence_passport_combined/requirements.txt
flask
paddleocr
paddlepaddle
python-docx
werkzeug
Pillow
